Update LF simulation to use 3HSP catalog
Updated simulation code as used for COSPAR 2021 and my thesis.
Now using the 3HSP catalog instead of 2WHSP for source selection.
Made minor improvements to appearance of plots.
Thinned the MCMC chains to increase independence of the samples.
Fixed bug in which the upper and lower errors on the parameters were
reversed.
Fixed bug in K-correction in which the photon index was used instead of
energy index.