Clean up code
Unused and redundant code has been removed.
Fixed bug in which the same mean exposure and index is used for all
samples, instead of being calculated separately for each sample.
Fixed bug in which the density errors were accidentally being set to the
density values in the 1/V_max calculation.